service fixed

+++ src/article/service/articleService.js
@@ -1,35 +1,34 @@
 import { cacheTime, myCache } from "../../../configuration/cache.config.js";
 import { nanoid } from "nanoid";
+import { storeToFirebase, deleteFromFirebase } from "../../../lib/storeFirebase.js";
 
 class ArticleService {
   constructor(database) {
     this.database = database;
   }
 
-  async createArticleService({
-    title,
-    post,
-    featured,
-    mainfeatured,
-    author,
-    category,
-    image,
-    imagename,
-  }) {
+  async createArticleService(articleData) {
     try {
-      const articleData = {
-        id: nanoid(),
-        title,
-        author,
-        post,
-        featured,
-        mainfeatured,
-        category,
-        image,
-        imagename,
-      };
-      const article = await this.database.createArticle(articleData);
-      return article;
+      const uploaded = await storeToFirebase(jobDetails.image);
+      const imageUrl = uploaded.imageURL;
+      const imageName = uploaded.imageName;
+      if (!uploaded.error) {
+        const articleContent = {
+          id: nanoid(),
+          image: imageUrl,
+          imagename: imageName,
+          title: articleData.title,
+          author: articleData.author,
+          post: articleData.post,
+          featured: articleData.post,
+          mainfeatured: articleData.featured,
+          category: articleData.category,
+        };
+        const article = await this.database.createArticle(articleContent);
+        return article;
+      } else {
+        return { error: uploaded.error };
+      }
     } catch (error) {
       console.error("service {create article}:", error.message);
     }
@@ -75,9 +74,7 @@ class ArticleService {
     try {
       const data = await myCache.get(cacheKey);
       if (data) return data;
-      const article = await this.database.readArticleByCategory(
-        articleCategory
-      );
+      const article = await this.database.readArticleByCategory(articleCategory);
       myCache.set(cacheKey, article, cacheTime);
       return article;
     } catch (error) {
@@ -120,29 +117,54 @@ class ArticleService {
     }
   }
 
-  async updateArticleService({
-    id,
-    title,
-    post,
-    featured,
-    mainfeatured,
-    category,
-    image,
-    imagename,
-  }) {
+  async updateArticleService(articleData) {
     try {
-      const articleData = {
-        id,
-        title,
-        post,
-        featured,
-        mainfeatured,
-        category,
-        image,
-        imagename,
-      };
-      const article = await this.database.updateArticle(articleData);
-      return article;
+      // admin wants to update the old image
+      if (articleData.image !== undefined) {
+        const deleteImage = await this.readArticleByIdService(articleData.id);
+        const deleteImageName = deleteImage.imagename;
+        const deletedImageFromFirebase = await deleteFromFirebase(deleteImageName);
+        if (deletedImageFromFirebase) {
+          const uploaded = await storeToFirebase(articleData.image);
+          const imageUrl = uploaded.imageURL;
+          const imageName = uploaded.imageName;
+          if (!uploaded.error) {
+            const articleContent = {
+              id: articleData.id,
+              image: imageUrl,
+              imagename: imageName,
+              title: articleData.title,
+              post: articleData.post,
+              featured: articleData.featured,
+              mainfeatured: articleData.mainfeatured,
+              category: articleData.category,
+            };
+            const article = await this.database.updateArticle(articleContent);
+            return article;
+          } else {
+            return { error: uploaded.error };
+          }
+        } else {
+          return { error: "Image wasn't deleted from firebase" };
+        }
+      } else {
+        // admin wants to keep the old image
+        const retrieveOldImage = await this.readArticleByIdService(id);
+        const retrievedOldImageName = retrieveOldImage.imagename;
+        const retrievedImageNameLink = retrieveOldImage.image;
+        const articleContent = {
+          id: articleData.id,
+          image: retrievedImageNameLink,
+          imagename: retrievedOldImageName,
+          title: articleData.title,
+          post: articleData.post,
+          featured: articleData.featured,
+          mainfeatured: articleData.mainfeatured,
+          category: articleData.category,
+        };
+        const article = await this.database.updateArticle(articleContent);
+        return article;
+      }
     } catch (error) {
       console.error("service {update article}", error.message);
     }
@@ -150,8 +172,15 @@ class ArticleService {
 
   async deleteArticleService(id) {
     try {
-      const article = await this.database.deleteArticle(id);
-      return article;
+      const getImage = await this.readArticleByIdService(id);
+      const image = getImage.imagename;
+      const deletedImageFromFirebase = await deleteFromFirebase(image);
+      if (deletedImageFromFirebase) {
+        const article = await this.database.deleteArticle(id);
+        return article;
+      } else {
+        return { error: "Could Not Delete Image From Firebase" };
+      }
     } catch (error) {
       console.error("service {delete article}", error.message);
     }
